+++ src/logdensityfunction.jl
@@ -30,8 +30,9 @@ using Random: Random
 """
     DynamicPPL.LogDensityFunction(
         model::Model,
-        getlogdensity::Function=getlogjoint_internal,
-        varinfo::AbstractVarInfo=VarInfo(model);
+        getlogdensity::Any=getlogjoint_internal,
+        varinfo::AbstractVarInfo=VarInfo(model)
+        accs::Union{NTuple{<:Any,AbstractAccumulator},AccumulatorTuple}=DynamicPPL.ldf_accs(getlogdensity);
         adtype::Union{ADTypes.AbstractADType,Nothing}=nothing,
     )
 
@@ -45,7 +46,9 @@ using `LogDensityProblems.logdensity` and `LogDensityProblems.logdensity_and_gra
 `adtype` is nothing, then only `logdensity` is implemented. If `adtype` is a concrete AD
 backend type, then `logdensity_and_gradient` is also implemented.
 
-There are several options for `getlogdensity` that are 'supported' out of the box:
+`getlogdensity` should be a callable which takes a single argument: a `VarInfo`, and returns
+a `Real` corresponding to the log density of interest. There are several functions in
+DynamicPPL that are 'supported' out of the box:
 
 - [`getlogjoint_internal`](@ref): calculate the log joint, including the log-Jacobian term
   for any variables that have been linked in the provided VarInfo.
@@ -67,6 +70,12 @@ There are several options for `getlogdensity` that are 'supported' out of the bo
 If you provide one of these functions, a `VarInfo` will be automatically created for you. If
 you provide a different function, you have to manually create a VarInfo and pass it as the
 third argument.
+
+`accs` allows you to specify an `AccumulatorTuple` or a tuple of `AbstractAccumulator`s
+which will be used _when evaluating the log density_`. (Note that the accumulators from the
+`VarInfo` argument are discarded.) By default, this uses an internal function,
+`DynamicPPL.ldf_accs`, which attempts to choose an appropriate set of accumulators based on
+which kind of log-density is being calculated.
 
 If the `adtype` keyword argument is provided, then this struct will also store the adtype
 along with other information for efficient calculation of the gradient of the log density.
@@ -139,11 +148,12 @@ struct LogDensityFunction{
     Tlink,
     M<:Model,
     AD<:Union{ADTypes.AbstractADType,Nothing},
-    F<:Function,
+    F,
     VNT<:VarNamedTuple,
     ADP<:Union{Nothing,DI.GradientPrep},
     # type of the vector passed to logdensity functions
     X<:AbstractVector,
+    AC<:AccumulatorTuple,
 }
     model::M
     adtype::AD
@@ -151,11 +161,19 @@ struct LogDensityFunction{
     _varname_ranges::VNT
     _adprep::ADP
     _dim::Int
+    _accs::AC
 
     function LogDensityFunction(
         model::Model,
-        getlogdensity::Function=getlogjoint_internal,
-        varinfo::AbstractVarInfo=VarInfo(model);
+        getlogdensity::Any=getlogjoint_internal,
+        # TODO(penelopeysm): It is a bit redundant to pass a VarInfo, as well as the
+        # accumulators, into here. The truth is that the VarInfo is used ONLY for generating
+        # the ranges and link status, so arguably we should only pass in a metadata; or when
+        # VNT is done, we should pass in only a VNT.
+        varinfo::AbstractVarInfo=VarInfo(model),
+        accs::Union{NTuple{<:Any,AbstractAccumulator},AccumulatorTuple}=ldf_accs(
+            getlogdensity
+        );
         adtype::Union{ADTypes.AbstractADType,Nothing}=nothing,
     )
         # Figure out which variable corresponds to which index, and
@@ -175,15 +193,26 @@ struct LogDensityFunction{
         end
         x = [val for val in varinfo[:]]
         dim = length(x)
+        # convert to AccumulatorTuple if needed
+        accs = AccumulatorTuple(accs)
         # Do AD prep if needed
         prep = if adtype === nothing
             nothing
         else
             # Make backend-specific tweaks to the adtype
             adtype = DynamicPPL.tweak_adtype(adtype, model, varinfo)
-            DI.prepare_gradient(
-                LogDensityAt{Tlink}(model, getlogdensity, all_ranges), adtype, x
-            )
+            args = (model, getlogdensity, all_ranges, accs)
+            if _use_closure(adtype)
+                DI.prepare_gradient(LogDensityAt{Tlink}(args...), adtype, x)
+            else
+                DI.prepare_gradient(
+                    logdensity_at,
+                    adtype,
+                    x,
+                    DI.Constant(Val{Tlink}()),
+                    map(DI.Constant, args)...,
+                )
+            end
         end
         return new{
             Tlink,
@@ -193,8 +222,9 @@ struct LogDensityFunction{
             typeof(all_ranges),
             typeof(prep),
             typeof(x),
+            typeof(accs),
         }(
-            model, adtype, getlogdensity, all_ranges, prep, dim
+            model, adtype, getlogdensity, all_ranges, prep, dim, accs
         )
     end
 end
@@ -207,12 +237,12 @@ end
 # LogDensityProblems.jl interface #
 ###################################
 """
-    ldf_accs(getlogdensity::Function)
+    ldf_accs(getlogdensity::Any)
 
 Determine which accumulators are needed for fast evaluation with the given
-`getlogdensity` function.
+`getlogdensity` callable.
 """
-ldf_accs(::Function) = default_accumulators()
+ldf_accs(::Any) = default_accumulators()
 ldf_accs(::typeof(getlogjoint_internal)) = default_accumulators()
 function ldf_accs(::typeof(getlogjoint))
     return AccumulatorTuple((LogPriorAccumulator(), LogLikelihoodAccumulator()))
@@ -223,40 +253,97 @@ end
 ldf_accs(::typeof(getlogprior)) = AccumulatorTuple((LogPriorAccumulator(),))
 ldf_accs(::typeof(getloglikelihood)) = AccumulatorTuple((LogLikelihoodAccumulator(),))
 
-struct LogDensityAt{Tlink,M<:Model,F<:Function,VNT<:VarNamedTuple}
+"""
+    logdensity_at(
+        params::AbstractVector{<:Real},
+        ::Val{Tlink},
+        model::Model,
+        getlogdensity::Any,
+        varname_ranges::VarNamedTuple,
+    ) where {Tlink}
+
+Calculate the log density at the given `params`, using the provided
+information extracted from a `LogDensityFunction`.
+"""
+function logdensity_at(
+    params::AbstractVector{<:Real},
+    ::Val{Tlink},
+    model::Model,
+    getlogdensity::Any,
+    varname_ranges::VarNamedTuple,
+    accs::AccumulatorTuple,
+) where {Tlink}
+    strategy = InitFromParams(VectorWithRanges{Tlink}(varname_ranges, params), nothing)
+    _, vi = DynamicPPL.init!!(model, OnlyAccsVarInfo(accs), strategy)
+    return getlogdensity(vi)
+end
+
+"""
+    LogDensityAt{Tlink}(
+        model::Model,
+        getlogdensity::Any,
+        varname_ranges::VarNamedTuple,
+        accs::AccumulatorTuple,
+    ) where {Tlink}
+
+A callable struct that behaves in the same way as `logdensity_at`, but stores the model and
+other information internally. Having two separate functions/structs allows for better
+performance with AD backends.
+"""
+struct LogDensityAt{Tlink,M<:Model,F,V<:VarNamedTuple,A<:AccumulatorTuple}
     model::M
     getlogdensity::F
-    varname_ranges::VNT
+    varname_ranges::V
+    accs::A
 
     function LogDensityAt{Tlink}(
-        model::M, getlogdensity::F, varname_ranges::N
-    ) where {Tlink,M,F,N}
-        return new{Tlink,M,F,N}(model, getlogdensity, varname_ranges)
+        model::M, getlogdensity::F, varname_ranges::V, accs::A
+    ) where {Tlink,M,F,V,A}
+        return new{Tlink,M,F,V,A}(model, getlogdensity, varname_ranges, accs)
     end
 end
 function (f::LogDensityAt{Tlink})(params::AbstractVector{<:Real}) where {Tlink}
-    strategy = InitFromParams(VectorWithRanges{Tlink}(f.varname_ranges, params), nothing)
-    accs = ldf_accs(f.getlogdensity)
-    _, vi = DynamicPPL.init!!(f.model, OnlyAccsVarInfo(accs), strategy)
-    return f.getlogdensity(vi)
+    return logdensity_at(
+        params, Val{Tlink}(), f.model, f.getlogdensity, f.varname_ranges, f.accs
+    )
 end
 
 function LogDensityProblems.logdensity(
     ldf::LogDensityFunction{Tlink}, params::AbstractVector{<:Real}
 ) where {Tlink}
-    return LogDensityAt{Tlink}(ldf.model, ldf._getlogdensity, ldf._varname_ranges)(params)
+    return logdensity_at(
+        params, Val{Tlink}(), ldf.model, ldf._getlogdensity, ldf._varname_ranges, ldf._accs
+    )
 end
 
 function LogDensityProblems.logdensity_and_gradient(
     ldf::LogDensityFunction{Tlink}, params::AbstractVector{<:Real}
 ) where {Tlink}
+    # `params` has to be converted to the same vector type that was used for AD preparation,
+    # otherwise the preparation will not be valid.
     params = convert(_get_input_vector_type(ldf), params)
-    return DI.value_and_gradient(
-        LogDensityAt{Tlink}(ldf.model, ldf._getlogdensity, ldf._varname_ranges),
-        ldf._adprep,
-        ldf.adtype,
-        params,
-    )
+    return if _use_closure(ldf.adtype)
+        DI.value_and_gradient(
+            LogDensityAt{Tlink}(
+                ldf.model, ldf._getlogdensity, ldf._varname_ranges, ldf._accs
+            ),
+            ldf._adprep,
+            ldf.adtype,
+            params,
+        )
+    else
+        DI.value_and_gradient(
+            logdensity_at,
+            ldf._adprep,
+            ldf.adtype,
+            params,
+            DI.Constant(Val{Tlink}()),
+            DI.Constant(ldf.model),
+            DI.Constant(ldf._getlogdensity),
+            DI.Constant(ldf._varname_ranges),
+            DI.Constant(ldf._accs),
+        )
+    end
 end
 
 function LogDensityProblems.capabilities(
@@ -289,6 +376,43 @@ model.
 By default, this just returns the input unchanged.
 """
 tweak_adtype(adtype::ADTypes.AbstractADType, ::Model, ::AbstractVarInfo) = adtype
+
+"""
+    _use_closure(adtype::ADTypes.AbstractADType)
+
+In LogDensityProblems, we want to calculate the derivative of `logdensity(f, x)` with
+respect to x, where f is the model (in our case LogDensityFunction or its arguments ) and is
+a constant. However, DifferentiationInterface generally expects a single-argument function
+g(x) to differentiate.
+
+There are two ways of dealing with this:
+
+1. Construct a closure over the model, i.e. let g = Base.Fix1(logdensity, f)
+
+2. Use a constant DI.Context. This lets us pass a two-argument function to DI, as long as we
+   also give it the 'inactive argument' (i.e. the model) wrapped in `DI.Constant`.
+
+The relative performance of the two approaches, however, depends on the AD backend used.
+Some benchmarks are provided here: https://github.com/TuringLang/DynamicPPL.jl/pull/1172
+
+This function is used to determine whether a given AD backend should use a closure or a
+constant. If `use_closure(adtype)` returns `true`, then the closure approach will be used.
+By default, this function returns `false`, i.e. the constant approach will be used.
+"""
+# For these AD backends both closure and no closure work, but it is just faster to not use a
+# closure (see link in the docstring).
+_use_closure(::ADTypes.AutoForwardDiff) = false
+_use_closure(::ADTypes.AutoMooncake) = false
+_use_closure(::ADTypes.AutoMooncakeForward) = false
+# For ReverseDiff, with the compiled tape, you _must_ use a closure because otherwise with
+# DI.Constant arguments the tape will always be recompiled upon each call to
+# value_and_gradient. For non-compiled ReverseDiff, it is faster to not use a closure.
+_use_closure(::ADTypes.AutoReverseDiff{compile}) where {compile} = !compile
+# For AutoEnzyme it allows us to avoid setting function_annotation
+_use_closure(::ADTypes.AutoEnzyme) = false
+# Since for most backends it's faster to not use a closure, we set that as the default
+# for unknown AD backends
+_use_closure(::ADTypes.AbstractADType) = false
 
 ######################################################
 # Helper functions to extract ranges and link status #